+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `CriteriaFilterParser` advanced filter inspired by OData semantic
 * `Annotation\Attribute::of` property, which is used for declare type of array items. 
 This will be used primary for OpenAPI schema
+* `ClassMetadata::getShortClassName` return class short name
 
 ### Changed
 * Now we insist that Relationships collections have to be `Doctrine\Common\Collections\Collection` 
@@ -18,6 +19,7 @@ This will be used primary for OpenAPI schema
 * Determination of attribute data type is now resolved in sequence:
     1. Try get type from return type of function
     2. Get type from setter parameter, if setter exists
+* `ClassMetadata::__construct` now accept reflection of class as first argument, instead string className
 
 ### Deprecated